experiments
The current issue with the approach seems to be the captured variable inference bug.
Upon closer inspection of the actual macros in Base, the spaw mechanism seems to avoid
the construction of a closure all together.

The use of the spawn multithreading mechanism will most likely be the optimal way forward here.

There are benchmarks of the different locking mechanism and SpinLock seems to be the best for
the use case here: https://gist.github.com/sjkelly/c46fc16e9bd98d9f992aaa75cc471619